Revert "feat(admin,api): 1740 - permettre aux modérateurs de supprimer une cl…"
This reverts commit 807f1b0cdc57c2de5d51f10c0441ac4f8030a930.
