+++ admin/src/types.ts
@@ -1,5 +1,4 @@
-import { ROLES, CLE_COLORATION_LIST, CLE_FILIERE_LIST, CLE_GRADE_LIST, STATUS_CLASSE_LIST, STATUS_PHASE1_CLASSE_LIST } from "snu-lib";
-
+import { ROLES } from "@/utils";
 export type Young = { _id: string };
 
 export type BusLine = { _id: string };
