+++ api/src/__tests__/helpers/app.js
@@ -27,7 +27,6 @@ function getAppHelper() {
   app.use("/table-de-repartition", require("../../controllers/planDeTransport/table-de-repartition"));
   app.use("/ligne-to-point", require("../../controllers/planDeTransport/ligne-to-point"));
   app.use("/ligne-de-bus", require("../../controllers/planDeTransport/ligne-de-bus"));
-  app.use("/cle", require("../../controllers/cle"));
 
   return app;
 }
