+++ packages/lib/src/dto/index.ts
@@ -1,2 +1,1 @@
 export * from "./cohortDto";
-export * from "./classeDto";
